Fixed Logging error,

+++ code/src/Data_processing/data_processor.h
@@ -1,4 +1,5 @@
 #include "encoder.h"
+#include <zephyr/types.h>
 
 
 
@@ -7,4 +8,4 @@ int find_local_minima(int data_array[][3]);
 void send_data_delta(int8_t rssi, int index);
 void send_data_zigma(int8_t rssi, int index);
 int8_t get_average(int8_t list[]);
-int *get_data();
+int16_t *get_data();
